+++ tests/music_theory/transpose.cpp
@@ -22,7 +22,7 @@
 
 #include <gtest/gtest.h>
 
-#include "music_theory/transposer.hpp"
+#include "music_theory/music_theory.hpp"
  
  using namespace music_theory;
